+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## Unreleased
 
 - Added a collection of theme colors to be available in the editor.
-- Added alternative style for Media & Text block.
+- Added "hover" style for Media & Text block.
+- Added "hover full width" style for Media & Text block.
 - Updated padding and font size for Button block.
 - Removed border radius option for Button block.
 - Changed default color for Button block.